+++ CHANGELOG.md
@@ -1,11 +1,18 @@
 # Precise UI Changelog
 
+## 0.8.3
+
+- `Skeleton` component added
+- Added onBeforeClose event to Modal
+
 ## 0.8.2
 
 - Updated `date-fns` dependency.
 - Fixed `InteractiveList` examples
 - Fixed truncation of long text in the `InteractiveList` for IE 11
 - Fixed regular expression for Babel in Webpack
+- Fixed Datepicker controlled mode issues
+- Fixed hiding suggestion list of `Autocomplete` after selection for IE 11
 
 ## 0.8.1